Core: Move eraseIf to it's own file
